+++ src/components/ReasonText.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Reason } from "../core/types";
+import { Reason } from "../core/types/types";
 import { Reasons } from "../theorems/reasons";
 
 export interface ReasonTextProps {
