+++ src/core/geometryText.tsx
@@ -35,7 +35,7 @@ export const angleStr = (a: string) => (
     {a}
   </span>
 );
-// for long-form reasons
+
 export const strs = {
   angle: "\u29A3",
   congruent: "\u2245",
