+++ src/core/templates/EqualSegments.tsx
@@ -1,7 +1,8 @@
-import { congruent, segmentStr } from "../../core/geometryText";
-import { Content } from "../../core/objgraph";
-import { SVGModes, Obj } from "../../core/types";
-import { StepFocusProps, StepTextProps, linked } from "../utils";
+import { linked } from "../../theorems/utils";
+import { Content } from "../diagramContent";
+import { congruent, segmentStr } from "../geometryText";
+import { StepFocusProps, StepTextProps } from "../types/stepTypes";
+import { Obj, SVGModes } from "../types/types";
 
 export class EqualSegments {
   static additions = (
