+++ src/core/templates/Midpoint.tsx
@@ -1,36 +1,35 @@
-import { perpendicular, segmentStr, strs } from "../../core/geometryText";
-import { Content } from "../../core/objgraph";
-import { Obj, SVGModes } from "../../core/types";
-import { StepFocusProps, StepTextProps, linked } from "../utils";
+import { linked } from "../../theorems/utils";
+import { Content } from "../diagramContent";
+import { segmentStr } from "../geometryText";
+import { StepFocusProps, StepTextProps } from "../types/stepTypes";
+import { Obj, SVGModes } from "../types/types";
 import { EqualSegments } from "./EqualSegments";
 
-export class Perpendicular {
+export class Midpoint {
   static additions = (
     props: StepFocusProps,
-    perp: string,
+    pt: string,
     segs: [string, string],
     num?: number,
     s2Mode?: SVGModes,
     ptMode?: SVGModes
   ) => {
-    props.ctx.getSegment(perp).mode(props.frame, ptMode || props.mode);
-    props.ctx.getSegment(segs[0]).mode(props.frame, ptMode || props.mode);
-    props.ctx.getSegment(segs[1]).mode(props.frame, ptMode || props.mode);
+    props.ctx.getPoint(pt).mode(props.frame, ptMode || props.mode);
+    EqualSegments.additions(props, segs, num, s2Mode || props.mode);
   };
   static text = (
     props: StepTextProps,
     label: string,
     segs: [string, string],
-    perp: string,
+    pt: string,
     num?: number
   ) => {
-    const perp1 = props.ctx.getSegment(perp);
     const s1 = props.ctx.getSegment(segs[0]);
     const s2 = props.ctx.getSegment(segs[1]);
     return (
       <span>
-        {linked(perp, perp1)}
-        {perpendicular}
+        {linked(pt, props.ctx.getPoint(pt))}
+        {" is the midpoint of "}
         {linked(label, s1, [
           s2,
           props.ctx.getTick(s1, Obj.EqualLengthTick, {
@@ -49,22 +48,22 @@ export class Perpendicular {
     ctx: Content,
     label: string,
     segs: [string, string],
-    perp: string
+    pt: string
   ) => {
     return (
       <span>
-        {linked(perp, ctx.getSegment(perp))}
-        {perpendicular}
+        {linked(pt, ctx.getPoint(pt))}
+        {" is the midpoint of "}
         {linked(label, ctx.getSegment(segs[0]), [ctx.getSegment(segs[1])])}
       </span>
     );
   };
-  static staticText = (s1: string, s2: string) => {
+  static staticText = (p: string, s: string) => {
     return (
       <span>
-        {segmentStr(s1)}
-        {perpendicular}
-        {segmentStr(s2)}
+        {p}
+        {" is the midpoint of "}
+        {segmentStr(s)}
       </span>
     );
   };
