+++ src/theorems/checking/pc2.tsx
@@ -1,23 +1,22 @@
+import { Content } from "../../core/diagramContent";
 import { Point } from "../../core/geometry/Point";
 import { Triangle } from "../../core/geometry/Triangle";
 import { comma } from "../../core/geometryText";
-import { Content } from "../../core/objgraph";
-import { Obj, SVGModes, Vector } from "../../core/types";
-import { Reasons } from "../reasons";
-import { EqualRightAngles } from "../templates/EqualRightAngles";
-import { EqualSegments } from "../templates/EqualSegments";
-import { EqualTriangles } from "../templates/EqualTriangles";
-import { Reflexive } from "../templates/Reflexive";
-import { RightAngle } from "../templates/RightAngle";
-import { SAS, SASProps } from "../templates/SAS";
+import { EqualRightAngles } from "../../core/templates/EqualRightAngles";
+import { EqualSegments } from "../../core/templates/EqualSegments";
+import { EqualTriangles } from "../../core/templates/EqualTriangles";
+import { Reflexive } from "../../core/templates/Reflexive";
+import { RightAngle } from "../../core/templates/RightAngle";
+import { SAS, SASProps } from "../../core/templates/SAS";
 import {
-  LayoutProps,
   StepFocusProps,
   StepMeta,
   StepTextProps,
   StepUnfocusProps,
-  makeStepMeta,
-} from "../utils";
+} from "../../core/types/stepTypes";
+import { LayoutProps, Obj, SVGModes, Vector } from "../../core/types/types";
+import { Reasons } from "../reasons";
+import { makeStepMeta } from "../utils";
 
 export const baseContent = (labeledPoints: boolean, parentFrame?: string) => {
   const coords: Vector[][] = [
